Move overlay to module.
