+++ assets/js/modules/analytics-4/components/dashboard/AudienceSegmentation/AudienceSegmentationIntroductoryOverlayNotification.js
@@ -26,11 +26,11 @@ import { __ } from '@wordpress/i18n';
  */
 import { Button } from 'googlesitekit-components';
 import Data from 'googlesitekit-data';
-import AudienceIntroductoryGraphicDesktop from '../../../svg/graphics/audience-segmentation-introductory-graphic-desktop.svg';
-import AudienceIntroductoryGraphicMobile from '../../../svg/graphics/audience-segmentation-introductory-graphic-mobile.svg';
-import { CORE_UI } from '../../googlesitekit/datastore/ui/constants';
-import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
-import OverlayNotification from './OverlayNotification';
+import AudienceIntroductoryGraphicDesktop from '../../../../../../svg/graphics/audience-segmentation-introductory-graphic-desktop.svg';
+import AudienceIntroductoryGraphicMobile from '../../../../../../svg/graphics/audience-segmentation-introductory-graphic-mobile.svg';
+import OverlayNotification from '../../../../../components/OverlayNotification/OverlayNotification';
+import { CORE_UI } from '../../../../../googlesitekit/datastore/ui/constants';
+import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
 
 const { useSelect, useDispatch } = Data;
